improvement of drag n drop events

+++ public/js/directives/droppable.js
@@ -2,9 +2,13 @@
 
 angular.module('bs').directive('droppable', function(){
   return {
+    restrict: 'A',
     scope: {
-      drop: '&', //parent
-      dropper: '=' //bi-directional scope
+      drop      : '&', //parent
+      dragover  : '&',
+      dragenter : '&',
+      dragleave : '&',
+      dropper   : '=' //bi-directional scope
     },
     link: function(scope, element){
       var elem = element[0];
@@ -13,16 +17,85 @@ angular.module('bs').directive('droppable', function(){
         if(event.preventDefault) event.preventDefault(); // allows us to drop
         this.classList.add('over');
         event.dataTransfer.dropEffect = 'copy';
+
+        var types = event.dataTransfer.types;
+        var l = types.length;
+        var rx = /^battleship\/(\w+)$/;
+        var self = this;
+
+        for(var i=0; i<l; i++){
+          var match = rx.exec(types[i].toLowerCase());
+          if(!match) continue;
+
+          var item = document.getElementById(match[1]);
+          if(item === null) throw new Error(this.id + '.dragenter failed because target item does not exist');
+          // call the passed dragenter function
+          scope.$apply(function(scope){
+            var dragover = scope.dragover();
+            if(typeof dragover === 'function'){
+              dragover(item, self);
+            }
+          });
+
+          return false;
+        }
+        
         return false;
       }, false);
 
       elem.addEventListener('dragenter', function(event){
         this.classList.add('over');
+
+        var types = event.dataTransfer.types;
+        var l = types.length;
+        var rx = /^battleship\/(\w+)$/;
+        var self = this;
+
+        for(var i=0; i<l; i++){
+          var match = rx.exec(types[i].toLowerCase());
+          if(!match) continue;
+          
+          var item = document.getElementById(match[1]);
+          if(item === null) throw new Error(this.id + '.dragenter failed because target item does not exist');
+          // call the passed dragenter function
+          scope.$apply(function(scope){
+            var dragenter = scope.dragenter();
+            if(typeof dragenter === 'function'){
+              dragenter(item, self);
+            }
+          });
+          
+          return false;
+        }
+        
         return false;
       }, false);
       
       elem.addEventListener('dragleave', function(event){
         this.classList.remove('over');
+
+        var types = event.dataTransfer.types;
+        var l = types.length;
+        var rx = /^battleship\/(\w+)$/;
+        var self = this;
+
+        for(var i=0; i<l; i++){
+          var match = rx.exec(types[i].toLowerCase());
+          if(!match) continue;
+
+          var item = document.getElementById(match[1]);
+          if(item === null) throw new Error(this.id + '.dragleave failed because target item does not exist');
+          // call the passed dragleave function
+          scope.$apply(function(scope){
+            var dragleave = scope.dragleave();
+            if(typeof dragleave === 'function'){
+              dragleave(item, self);
+            }
+          });
+
+          return false;
+        }
+        
         return false;
       }, false);
 
@@ -30,13 +103,14 @@ angular.module('bs').directive('droppable', function(){
         if(event.stopPropagation) event.stopPropagation(); // stops some browser from redirecting
         this.classList.remove('over');
         var item = document.getElementById(event.dataTransfer.getData('Text'));
+        if(item === null) throw new Error(elem.id + '.drop failed because target item does not exist');
         item.parentNode.removeChild(item);
-        var dropId = this.id;
+        var self = this;
         // call the passed drop function
         scope.$apply(function(scope){
-          var fn = scope.drop();
-          if(typeof fn !== 'undefined'){
-            fn(item.id, dropId);
+          var drop = scope.drop();
+          if(typeof drop === 'function'){
+            drop(item, self);
           }
         });
